OPABackend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cfgv==3.3.1
distlib==0.3.4
Django==4.0.5
django-allauth==0.50.0
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.13.1
djongo==1.3.6
dj-rest-auth==2.2.4
drf-yasg==1.20.0
filelock==3.7.1
identify==2.5.1
nodeenv==1.6.0
platformdirs==2.5.2
pre-commit==2.19.0
pymongo==3.12.1
pytz==2022.1
PyYAML==6.0
toml==0.10.2
virtualenv==20.14.1
